Update to use Flask-JWT-Extended - Using claims to check if user is admin

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from models.item import ItemModel
 
@@ -62,6 +62,10 @@ class Item(Resource):
     
     @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        print (f"JWT claims: {claims}")
+        if not claims['is_admin']:
+            return {"message": "Admin privilege required"}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
